backend/requirements.txt
fastapi
uvicorn
pydantic-settings
openai
pandas
python-docx
pypdf
python-multipart
httpx
pyyaml
sqlalchemy
pymysql
cryptography
alembic
aiofiles
python-jose
neo4j
